have added the change() method to TrafficLight class. using a random number and a switch statement the lights will change to one of 4 settings. Additionally updatePosition() in Car class has been updated to handel when lights are red and green

+++ src/Road.java
@@ -1,7 +1,7 @@
 public class Road {
-    public int length;
-    public int xStart,yStart;
-    public int xFinish,yFinish;
+    public double length;
+    public double xStart,yStart;
+    public double xFinish,yFinish;
     public String orientation;
 
     public Road(int length, int x, int y, String orientation) {
